+++ b/api-tlahtolli-admin-dasboard/src/main/java/com/tlahtolli/api/controller/UserAchievementController.java
@@ -23,7 +23,7 @@ public class UserAchievementController {
 	 * del frontend.
 	 */
 	@GetMapping
-	public ResponseEntity<List<AchievementWithStatus>> getAll(@RequestParam Long userId) {
+	public ResponseEntity<List<AchievementWithStatus>> getAll(@RequestParam Integer userId) {
 		return ResponseEntity.ok(achievementService.getAllWithStatus(userId));
 	}
 }
