+++ b/api-tlahtolli-admin-dasboard/src/main/java/com/tlahtolli/api/dto/response/LessonResponse.java
@@ -3,12 +3,12 @@ package com.tlahtolli.api.dto.response;
 import com.tlahtolli.api.entity.Lesson;
 
 public record LessonResponse(
-        Long id,
-        Long unitId,
-        Integer lessonNum,
+        Integer id,
+        Integer unitId,
+        Short lessonNum,
         String title,
         String description,
-        Integer isFree,
+        Short isFree,
         Integer xpReward
 ) {
     public static LessonResponse from(Lesson l) {
